feat: add branching to workflows and merge

+++ nodecosmos/src/api/like_api.rs
@@ -1,32 +1,27 @@
-use crate::api::data::RequestData;
-use crate::api::types::Response;
-use crate::models::like::Like;
-use crate::models::materialized_views::likes_by_user::LikesByUser;
-use crate::models::user::CurrentUser;
 use actix_web::{delete, get, post, web, HttpResponse};
 use charybdis::operations::{DeleteWithCallbacks, Find, InsertWithCallbacks};
 use scylla::CachingSession;
 use serde_json::json;
 
-#[get("/{objectId}/{branchId}")]
-pub async fn get_like_count(db_session: web::Data<CachingSession>, like: web::Path<Like>) -> Response {
-    let like_count = match like.find_by_primary_key().execute(&db_session).await.ok() {
-        Some(mut like) => like.like_count(&db_session).await?,
-        None => 0,
-    };
+use crate::api::data::RequestData;
+use crate::api::types::Response;
+use crate::models::like::Like;
+use crate::models::materialized_views::likes_by_user::LikesByUser;
+use crate::models::user::CurrentUser;
+
+#[get("/{objectId}/{branchId}/{objectType}")]
+pub async fn get_like_count(db_session: web::Data<CachingSession>, mut like: web::Path<Like>) -> Response {
+    let like_count = like.like_count(&db_session).await?;
 
     Ok(HttpResponse::Ok().json(json!({
         "id": like.object_id,
         "branchId": like.branch_id,
-        "likesCount": like_count,
+        "likeCount": like_count,
     })))
 }
 
 #[post("")]
 pub async fn create_like(data: RequestData, mut like: web::Json<Like>) -> Response {
-    like.user_id = data.current_user.id;
-    like.username = data.current_user.username.clone();
-
     like.insert_cb(&data).execute(data.db_session()).await?;
 
     let like_count = like.like_count(data.db_session()).await?;
@@ -34,14 +29,17 @@ pub async fn create_like(data: RequestData, mut like: web::Json<Like>) -> Respon
     Ok(HttpResponse::Ok().json(json!({
         "id": like.object_id,
         "branchId": like.branch_id,
-        "likesCount": like_count,
+        "likeCount": like_count,
     })))
 }
 
 #[delete("/{objectId}/{branchId}")]
-pub async fn delete_like(data: RequestData, mut like: web::Path<Like>) -> Response {
-    like.user_id = data.current_user_id();
+pub async fn delete_like(data: RequestData, like: web::Path<Like>) -> Response {
     let mut like = like.find_by_primary_key().execute(data.db_session()).await?;
+
+    if like.user_id != data.current_user_id() {
+        return Ok(HttpResponse::Forbidden().finish());
+    }
 
     like.delete_cb(&data).execute(data.db_session()).await?;
 
@@ -50,7 +48,7 @@ pub async fn delete_like(data: RequestData, mut like: web::Path<Like>) -> Respon
     Ok(HttpResponse::Ok().json(json!({
         "id": like.object_id,
         "branchId": like.branch_id,
-        "likesCount": like_count,
+        "likeCount": like_count,
     })))
 }
